requirements.txt
requests
openai
schedule
python-telegram-bot
pytz